# verity-server/spring-boot-server/springServer/src/main/resources/verityCore-swagger20.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.1",
        "title": "VerityReputation",
...
    },
    "host": "localhost:8080",
    "basePath": "/verity/core",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Basic Auth": {
            "type": "basic",
            "description": "Basic Authentication for now while we build and test the api"
        },
        "verity_auth_token": {
            "name": "verity_auth_token",
            "type": "apiKey",
            "in": "query"
        }
    },
    "paths": {
...
        "/transaction": {
            "post": {
                "operationId": "createTransaction",
                "summary": "create transaction",
                "tags": [
                    "Transaction"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "hash - Ethereum address of the transaction. For demo / proof of concept this will be a db id or offchain hash.",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    },
                    "401": {
                        "$ref": "#/responses/trait:standardErrors:401"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standardErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standardErrors:404"
                    },
                    "409": {
                        "$ref": "#/responses/trait:standardErrors:409"
                    },
                    "422": {
                        "$ref": "#/responses/trait:standardErrors:422"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standardErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "createTransaction",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            }
        },
        "/transaction/{uuid}": {
            "parameters": [
                {
                    "name": "uuid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "getTransaction",
                "summary": "get transaction",
                "tags": [
                    "Transaction"
                ],
                "description": "Get a specific transaction by id",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    },
                    "401": {
                        "$ref": "#/responses/trait:standardErrors:401"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standardErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standardErrors:404"
                    },
                    "409": {
                        "$ref": "#/responses/trait:standardErrors:409"
                    },
                    "422": {
                        "$ref": "#/responses/trait:standardErrors:422"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standardErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "getTransaction",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            }
        },
        "/transaction/agent/{agentId}": {
            "parameters": [
                {
                    "name": "agentId",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "getTransactionsByAgent",
                "summary": "get Transactions by AgentId",
                "tags": [
                    "Transaction"
                ],
                "description": "Get all transaction for an agent. Use query parameters to limit results. Transactions will be returned where the agent is both the source or the target unless limited by the 'direction' parameter.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "limit by source or target, e.g. the the agent is sending or receving the transaction.",
                        "type": "string",
                        "enum": [
                            "source",
                            "target"
                        ]
                    },
                    {
                        "$ref": "#/parameters/trait:paged:pageNumber"
                    },
                    {
                        "$ref": "#/parameters/trait:paged:pageSize"
                    },
                    {
                        "$ref": "#/parameters/trait:byDateRange:startDate"
                    },
                    {
                        "$ref": "#/parameters/trait:byDateRange:endDate"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Transaction"
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/responses/trait:standardErrors:401"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standardErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standardErrors:404"
                    },
                    "409": {
                        "$ref": "#/responses/trait:standardErrors:409"
                    },
                    "422": {
                        "$ref": "#/responses/trait:standardErrors:422"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standardErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "getTransactionsByAgent",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            }
        },
...
    },
    "parameters": {
...
        "trait:paged:pageNumber": {
            "name": "pageNumber",
            "in": "query",
            "description": "starting page number. Must be > 0.",
            "required": true,
            "type": "integer",
            "minimum": 1
        },
        "trait:paged:pageSize": {
            "name": "pageSize",
            "in": "query",
            "description": "number of records to return per page.",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
        },
        "trait:byDateRange:startDate": {
            "name": "startDate",
            "in": "query",
            "required": true,
            "type": "string"
        },
        "trait:byDateRange:endDate": {
            "name": "endDate",
            "in": "query",
            "required": true,
            "type": "string"
        }
    },
    "responses": {
        "trait:standardErrors:401": {
            "description": "",
            "schema": {
                "type": "object"
            }
        },
        "trait:standardErrors:403": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            }
        },
        "trait:standardErrors:404": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "status",
                    "error"
                ]
            }
        },
        "trait:standardErrors:409": {
            "description": "",
            "schema": {
                "type": "object",
                "description": "409 Conflict. https://tools.ietf.org/html/rfc7231#section-6.5.8"
            }
        },
        "trait:standardErrors:422": {
            "description": "",
            "schema": {
                "type": "object",
                "description": "422 Unprocessable Entity. http://tools.ietf.org/html/rfc4918#section-11.2"
            }
        },
        "trait:standardErrors:500": {
            "description": "",
            "schema": {
                "type": "object"
            }
        }
    },
    "definitions": {
...
        "Transaction": {
            "title": "Transaction",
            "type": "object",
            "description": "A fundamental unit of reputation value transfer, attestation or claim. Represents a singular event that transfers or bestows reputation to the target/recipient. The sender/source can be a person, software agent, organization or the like. Intended to be independent of database or blockchain so all IDs should be globally identifiable hashes in multi-hash format pointing to the canonical representation or permanent public store.\n\nSimilar to and should map to TrustAtom https://github.com/CoMakery/trust-exchange/blob/master/README.md",
            "properties": {
                "uuid": {
                    "type": "string",
                    "description": "UUID, GUID, HASH,  MultiHash or ProxyContract Address that represents this object"
                },
                "valueActionId": {
                    "type": "string",
                    "description": "Hash link to description of how the score will be derived, or the community feature that affects the way the value will be interpreted. Could be a globally unique link or content addressable graph pointing to a richer source of description or information."
                },
                "sourceAgentId": {
                    "type": "string"
                },
                "targetAgentId": {
                    "type": "string"
                },
                "timeStamp": {
                    "type": "string",
                    "description": "timestamp or block height of blockchain at time of creation. Should be self-describing as to format and meaning."
                },
                "value": {
                    "format": "string",
                    "description": "value (points) awarded at the time of the transaction. Immutable. Use ValueActionId to lookup the 'current' value if you want to calculate points based on the latest default value as apposed to the value awarded at time of the transaction. This allows for 'retroactive' adjustment of points.",
                    "type": "string"
                }
            },
            "required": [
                "uuid",
                "valueActionId",
                "sourceAgentId",
                "targetAgentId",
                "timeStamp"
            ],
            "x-stoplight": {
                "id": "Transaction",
                "name": "Transaction",
                "public": true
            }
        },
...
    },
...
}